refact: reorganize redux files
