+++ src/components/bank-accounts/list/destroy.jsx
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { withRouter, Link } from 'react-router-dom';
 
 import { addFlashMessage } from 'actions/flash-messages.js';
-import { deleteBankAccount } from 'actions/bank-accounts.js';
+import { deleteBankAccount } from 'services/bank-accounts';
 import { getCurrentOrganizationId } from 'selectors/organizations.js';
 import { confirm } from 'components/utils/confirm';
