+++ src/services/bank-accounts/types.ts
@@ -1,0 +1,8 @@
+import { Money } from 'utils/money';
+
+export interface BankAccount {
+  id:       number;
+  name:     string;
+  currency: string;
+  balance:  Money;
+}
