--- src/utils/bank-account/index.ts
@@ -1,1 +1,0 @@
-export * from './bank-account';
